app/bot/pyproject.toml
[project]
name = "rabbit-bot"
version = "0.1.0"
description = "Bot Rabbit BAR and others"
authors = [
    {name = "TitanPillow"}
]
readme = "README.md"
requires-python = ">=3.8.0"
dependencies = [
    "aiogram (>=3.4.1,<3.5.0)",
    "asyncpg (>=0.29.0,<0.30.0)",
    "qrcode (>=6.1.0,<7.1.0)",
    "pillow (>=10.1.0,<11.6.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"